requirements.txt
openai
python-dotenv
opencv-python
python-pptx 
PyPDF2
python-docx
google-adk
pandas
openpyxl
tqdm
Pillow
pymysql
jupyter
# pywin32
dashscope
pyyaml
litellm
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0
python-multipart>=0.0.5
openai>=1.0.0